requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.14
sqlparse==0.5.1
typing_extensions==4.12.2
django-mptt==0.13.4
django_extensions=3.2.3
django-filter==24.2
djangorestframework==3.15.2